backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
gunicorn==23.0.0
streamlit==1.41.1
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2
requests==2.32.3
beautifulsoup4==4.12.3
PyPDF2==3.0.1
pdfplumber==0.11.5
python-docx==1.1.2
spacy==3.8.2
